requirements.txt
streamlit
SpeechRecognition
pydub
google-generativeai
python-dotenv
# openai-whisper
